+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 - fix: 修复 webdav 同步时不兼容问题
 - feat: popup 帖子列表显示话题数
 - refactor: 其他代码优化
+- refactor: 引入 Arco Design UI，抛弃 Element UI